chore: props enhancement for Grid
